+++ backend/src/schemas/answer.ts
@@ -1,13 +1,13 @@
 import { z } from 'zod';
 
 export const createAnswerSchema = z.object({
-  content: z.string().min(5),
+  content: z.string().min(5).max(20000),
 });
 
 export type CreateAnswerBody = z.infer<typeof createAnswerSchema>;
 
 export const updateAnswerSchema = z.object({
-  content: z.string().min(5),
+  content: z.string().min(5).max(20000),
 });
 
 export type UpdateAnswerBody = z.infer<typeof updateAnswerSchema>;